Add proper loading screen
In most cases, the screen just appears and disappears in a blink. So it is worth considering if we should always keep it visible for a like 3 seconds, so you can actually see what is going on. Could also consider having just a blank screen and only show a proper loading screen after like 100-200 ms. But forcing a loading screen during development is pretty annoying. Maybe hide it behind a "Developer"-flag. This would also include the the "Dev Mode" button on the frontpage.