Recreate default team files for new client releases
This is mostly a development setting to prevent too many serialization errors. We are still iterating on the file- serialization format, so it happens often that de-serialization fails because the team files was of an older date. At least this should prevent it. We probably need to rethink this approach as it gets more clear how teams are stored and the serialization format matures.