Only send RerollSourceId across the network
RerollSources are not data classes, so we cannot compare them using equals on both sides of a network transfer. This commit refactores this so we are now only sending the IDs and then all participants can look them up themselves.
This problem caused Team rerolls to not be counted correctly in a P2P game, causing clients to get out of sync.

There is a also a bunch of formatting that was left out of previous commits.