Optimize copying when creating an UiSnapshot + fix display bug on WASM
It looked like something was going wrong on WASM with change detection when showing the Player Context Menu. Maybe I was doing something wrong and I was lucky it just worked on JVM, but looking into it, I realized that we ended up copying the UiFieldSquare quite a lot, which didn't seem strictly needed. So I changed it from a data class to a normal class but with mutableStateOf fields. I didn't measure the memory usage, but it fixed the error on WASM. The only downside is that the UIFieldSquare now has a lot of var properties, but that is probably fine as it exposes the fact that you end up modifying the square in the field decorators.

This change also fixed a bug with display move-used, not resetting correctly under Undo actions. That seems likely since that also relies on state inside the UiFieldSquare.