Rework Team and Player serialization + introduce SkillSettings class
This commit was motivated by the Team class starting to be really annoying to serialize which resulted in a lot of @Transient annotations and having to think about restoring cyclic references. So in this commit we introduce two new classes SerializedTeam and SerializedPlayer, which is a way to manually define how teams gets sent across the wire or stored on disk. This makes it more flexible and allows us to "dumb" the serialized version way down since we do not need all the temporary properties on the Team and Player models.

This also resulted in a larger refactor how Skills are created, so this commit also introduces a new SkillSettings class that is responsible for creating skills for a rulesets. This means all SkillFactory classes are now gone. This also created a more flexible approach where SkillId is now used to identify skills independent of rulesets and their exact behaviour is then defined by the SkillSetttings class.