Unify dialog setup
First step at making it easier to create new dialogs. This commit adds a new general JervisDialog that makes it easier to setup a "standard" dialog. It can probably be optimized further. Especially with regard to setting the icon, but that can come later.

Also fixed a few bugs with the FUMBBL login screen that didn't show error messages correctly.