Refactor player icon and portrait loading and replace roster logos
After talking to Christer, it looks like the licensing situation for roster logos, player icons and portraits are a bit unclear, so these cannot be included as part of the application bundle (and thus, this repository), but loading them dynamically at runtime is fine. So this is what this commit does. The `icons.ini` from the FUMBBL project has been copied and act as a map for finding all player icons and portraits. On first access they are downloaded and cached locally.

A few other changes has also been included:
- LICENSE and README has been updated to reflect this.
- All roster icons are also removed. New Jervis specific ones are replacing them, and the FUMBBL roster logos are only used when importing teams from there.
- A proxy has been added to http://jervis.ilios.dk which makes it possible to download player icons on WASM. They are currently protected by CORS.
- Renamed and reworked the resource folders so it is very clear which file are coming from FUMBBL and which are not.