Show players in Dogouts from the first frame
Using `combine` in the dogout flow had the side-effect that you needed two events to render players in the dogout. This meant they were not shown until the 2nd Fan Factor roll.
This bug was introduced with the removal of `Team.dogoutFlow`, but is fixed by only relying on the uiSnapshot flow instead.