Fix touchbacks being triggered if ball lands in No Man's Land
While this rule in theory only applies to to BB7. With customized fields, the situation could also occur in other rulesets, so it makes sense to use the Errata for this across rulesets.
So with this fix, a touchback is only awarded if the ball goes out of bounds or lands behind the kicking teams line of scrimmage. Before this fix, a touchback was triggered if the ball
landed across the half line (note, this is also a possible interpretation of the rules, but for now we are going with the entire NML zone).