Fix timer button not receiving all updates correctly
This was discovered on Web (it worked by accident on JVM). We need to check all places that use `collectAsState`. Updates from this only trigger when there is an object ref change. So either we need to emit immutable data classes in these places or classes with `mutableStateOf` properties. Right now we are using a mix of normal classes and mutableStateOf properties, but it is easy to forget when making changes to the codebase, so it would probably be good to find a. single approach that can be used everywhere.