Player Card now show keywords and player status labels below the skill list
For now, the colors have just been chosen to match the screenshots of markers shared by Games Workshop. They should be evaluated at some point.

The order of labels are:
1) Player State
2) Status indicators
3) Player keywords

This order is somewhat arbitrarily chosen, but the current order does mean that you only need to look at the front of the list to see the "important" labels.